Modify test to run in both connect and local
