+++ sql/core/src/test/scala/org/apache/spark/sql/connector/StaticProcedureSuite.scala
@@ -1,0 +1,29 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.sql.connector
+
+import org.apache.spark.sql.{SparkSessionBuilder, StaticProcedureSuiteBase}
+import org.apache.spark.sql.connector.catalog.InMemoryTableCatalog
+import org.apache.spark.sql.test.SharedSparkSession
+
+class StaticProcedureSuite extends SharedSparkSession with StaticProcedureSuiteBase {
+
+  override protected def configure(builder: SparkSessionBuilder): builder.type = {
+    builder.config(s"spark.sql.catalog.testcat", classOf[InMemoryTableCatalog].getName)
+  }
+}
